% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ios.R
\name{apple_ios_version_history}
\alias{apple_ios_version_history}
\title{Retrieve Apple iOS  Version Release History}
\usage{
apple_ios_version_history()
}
\description{
Reads \url{https://en.wikipedia.org/wiki/IOS_version_history"} to build a data
frame of Apple iOS version release numbers and dates with pseudo-semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
\note{
This does not distinguish by device target and only pulls the first release
date and excludes betas. If more granular data is needed, file an issue or PR.
}
