% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrix-netscaler.R
\name{citrix_netscaler_version_history}
\alias{citrix_netscaler_version_history}
\title{Citrix Netscaler ADC Version History}
\usage{
citrix_netscaler_version_history()
}
\description{
Reads \url{https://support.citrix.com/article/CTX121840} to build a data frame of
Citrix Netscaler ADC version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
