% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complete_semver}
\alias{complete_semver}
\title{Turn partial "valid" semantic version strings into a complete semver-tri or quad strings}
\usage{
complete_semver(x, quad = FALSE)
}
\arguments{
\item{x}{a character vector of full or partial version strings}

\item{quad}{(logical) if `TRUE` then a three-dot semver is returned, else a two-dot semver
is returned. Default: `FALSE`.}
}
\description{
For MAJOR.MINOR.PATCH (semver-tri), turn `1` into `1.0.0`; `1.1` into `1.1.0`; `1.1.1`
into `1.1.1`. For MAJOR.MINOR.PATCH.EXTENSION (semver-quad), turn `1` into `1.0.0.0`;
`1.1` into `1.1.0.0`; `1.1.1.0` into `1.1.1.0`.
}
\details{
Partial validity checking is performed to test if the input strings contain only
digits and periods. "Invalid" input is returned unscathed.
}
