% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etcd.R
\name{etcd_version_history}
\alias{etcd_version_history}
\title{Retrieve etcd Version Release History}
\usage{
etcd_version_history(refresh = FALSE)
}
\arguments{
\item{refresh}{if \code{TRUE} and there \verb{~/.vershist} cache dir exists, will
cause the version history database for apache to be rebuilt. Defaults
to \code{FALSE} and has no effect if \verb{~/.vershist} cache dir does not exist.}
}
\description{
Reads \url{https://github.com/etcd-io/etcd/releases} to build a data frame of
etcd version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
\note{
This function requires a valid GitHub API key stored in \code{GITHUB_PAT}
}
