% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exim.R
\name{exim_version_history}
\alias{exim_version_history}
\title{Retrieve Exim Version Release History}
\usage{
exim_version_history(refresh = FALSE)
}
\arguments{
\item{refresh}{if \code{TRUE} and there \verb{~/.vershist} cache dir exists, will
cause the version history database for apache to be rebuilt. Defaults
to \code{FALSE} and has no effect if \verb{~/.vershist} cache dir does not exist.}
}
\description{
Reads \url{https://marc.info/?l=exim-announce&r=1&w=2}
to build a data frame of Exim version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
