% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{google_chrome_version_history}
\alias{google_chrome_version_history}
\title{Retrieve Google Chrome Version Release History}
\usage{
google_chrome_version_history(refresh = FALSE)
}
\arguments{
\item{refresh}{if \code{TRUE} and there \verb{~/.vershist} cache dir exists, will
cause the version history database for apache to be rebuilt. Defaults
to \code{FALSE} and has no effect if \verb{~/.vershist} cache dir does not exist.}
}
\description{
Reads \url{https://en.wikipedia.org/wiki/Google_Chrome_version_history"} to build a data
frame of Google Chrome version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
\note{
This \emph{only} pulls the first release date and does not distinguish by OS.
If more granular data is needed, file an issue or PR.
}
