% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_valid}
\alias{is_valid}
\title{Test if semantic version strings are valid}
\usage{
is_valid(v)
}
\arguments{
\item{v}{character verctor of version strings}
}
\description{
Test if semantic version strings are valid
}
