% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_valid_semver}
\alias{is_valid_semver}
\title{Test if semantic version strings are valid tri-string}
\usage{
is_valid_semver(v)
}
\arguments{
\item{v}{character verctor of version strings}
}
\description{
Test if semantic version strings are valid tri-string
}
