% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memcached.R
\name{memcached_version_history}
\alias{memcached_version_history}
\title{Retrieve memcached Version Release History}
\usage{
memcached_version_history()
}
\description{
Clones the \code{memcached} wiki and reads \code{ReleaseNotes.md} to build a data frame of
\code{memcached} version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor. It uses \code{\link[git2r:cred_ssh_key]{git2r::cred_ssh_key()}} so you \emph{must} have SSH keys
setup in order for this function to work.
}
\note{
This is an \emph{expensive} function as it does quite a wiki clone.
Please consider using some sort of cache for the results unless
absolutely necessary.
}
