% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongodb.R
\name{mongodb_version_history}
\alias{mongodb_version_history}
\title{Retrive MongoDB Version Release History}
\usage{
mongodb_version_history()
}
\description{
Reads \url{https://www.mongodb.org/dl/linux"} to build a data frame of
MongoDB version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
\note{
Release candidates are not included and release history is only
supported for linux releases.
}
