% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nginx.R
\name{nginx_version_history}
\alias{nginx_version_history}
\title{Retrieve nginx Version Release History}
\usage{
nginx_version_history(refresh = FALSE)
}
\arguments{
\item{refresh}{if \code{TRUE} and there \code{~/.vershist} cache dir exists, will
cause the version history database for apache to be rebuilt. Defaults
to \code{FALSE} and has no effect if \code{~/.vershist} cache dir does not exist.}
}
\description{
Reads \url{https://raw.githubusercontent.com/nginx/nginx/master/docs/xml/nginx/changes.xml}
to build a data frame of nginx version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
