% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openssh.R
\name{openssh_version_history}
\alias{openssh_version_history}
\title{Retrieve OpenSSH (non-portable) Version Release History}
\usage{
openssh_version_history()
}
\description{
Reads \url{https://www.openssh.com/releasenotes.html} to build a data frame of
OpenSSH (non-portable) version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
\note{
File an issue or PR if the "portable" equivalent is needed
}
