% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/php.R
\name{php_version_history}
\alias{php_version_history}
\title{Retrieve PHP Version Release History}
\usage{
php_version_history()
}
\description{
Reads \url{https://secure.php.net/releases/} to build a data frame of PHP version
release numbers and dates with semantic version strings parsed and separate
fields added. The data frame is also arranged in order from lowest version
to latest version and the \code{vers} column is an ordered factor.
}
