% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendmail.R
\name{sendmail_version_history}
\alias{sendmail_version_history}
\title{Retrieve sendmail Version Release History}
\usage{
sendmail_version_history()
}
\description{
Reads \url{ftp://ftp.sendmail.org/pub/sendmail/"} to build a data frame of
\code{sendmail} version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
