% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_version_history}
\alias{sqlite_version_history}
\title{Retrieve SQLite Version Release History}
\usage{
sqlite_version_history()
}
\description{
Scrapes \url{https://sqlite.org/changes.html} to build a data frame of
openresty version release numbers and dates with pesudo-semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
