% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualbox.R
\name{virtualbox_version_history}
\alias{virtualbox_version_history}
\title{Retrieve Virtualbox Version Release History}
\usage{
virtualbox_version_history()
}
\description{
Reads \url{https://download.virtualbox.org/virtualbox/} to build a data frame of
Virtualbox version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
