% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache-httpd.R
\name{apache_httpd_version_history}
\alias{apache_httpd_version_history}
\title{Retrieve Apache httpd Version Release History}
\usage{
apache_httpd_version_history()
}
\description{
Reads \url{https://github.com/apache/httpd} releases to build a data frame of
Apache \code{httpd} version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
