% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etcd.R
\name{etcd_version_history}
\alias{etcd_version_history}
\title{Retrieve etcd Version Release History}
\usage{
etcd_version_history()
}
\description{
Reads \url{https://github.com/etcd-io/etcd/releases} to build a data frame of
etcd version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
\note{
This function requires a valid GitHub API key stored in \code{GITHUB_PAT}
}
