% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{mysql_version_history}
\alias{mysql_version_history}
\title{Retrieve MySQL Version Release History}
\usage{
mysql_version_history()
}
\description{
Scrapes \url{https://downloads.mysql.com/archives/community/} to build a data frame of
openresty version release numbers and dates with pesudo-semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
\details{
The selector for versioning is "\code{Generic Linux (Architecture Independent)}" and
the first found date is used for the \code{rls_date}. File an issue or PR if
alternate behaviour is required.
}
\note{
This is an \emph{expensive} function as it does quite a bit of scraping.
Please consider using some sort of cache for the results unless
absolutely necessary.
}
