% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nginx.R
\name{nginx_version_history}
\alias{nginx_version_history}
\title{Retrive nginx Version Release History}
\usage{
nginx_version_history()
}
\description{
Reads \url{https://raw.githubusercontent.com/nginx/nginx/master/docs/xml/nginx/changes.xml}
to build a data frame of nginx version release numbers and dates with semantic version
strings parsed and separate fields added. The data frame is also arranged in
order from lowest version to latest version and the \code{vers} column is an
ordered factor.
}
