% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vershist-package.R
\docType{package}
\name{vershist}
\alias{vershist}
\alias{vershist-package}
\title{Collect Version Histories For Vendor Products}
\description{
Provides a set of functions to gather version histories of products
(mainly software products) from their sources (generally websites).
}
\author{
Bob Rudis (bob@rud.is)
}
